requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# 任务队列
celery==5.3.4
flower==2.0.1

# HTTP客户端
httpx==0.26.0
requests==2.31.0

# 网页抓取
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.0

# AI/ML
openai==1.10.0
langchain==0.1.5
tiktoken==0.5.2

# 邮件
sendgrid==6.11.0
python-mailgun==0.1.1
brevo-python==1.0.0
jinja2==3.1.3

# 对象存储
minio==7.2.3

# 工具
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 数据处理
pandas==2.2.0
numpy==1.26.3

# 时间处理
python-dateutil==2.8.2
pytz==2024.1

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# 代码质量
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

